wip: finished precedence
