+++ crates/gdtk-ast/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2021"
 
 [dependencies]
 enum-as-inner = "0.6.0"
+prec = { version = "1.1.0", git = "https://github.com/mcpar-land/prec" }
